requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
pandas>=1.5.0
odfpy>=1.4.0
numpy
openai>=1.0.0
pandas>=1.5.0
openpyxl>=3.0.0
